+++ linker_test.go
@@ -2,6 +2,7 @@ package goa4web
 
 import (
 	"context"
+	"database/sql"
 	"net/http/httptest"
 	"testing"
 	"time"
@@ -9,6 +10,29 @@ import (
 	"github.com/DATA-DOG/go-sqlmock"
 )
 
+func TestLinkerFeed(t *testing.T) {
+	rows := []*GetAllLinkerItemsByCategoryIdWitherPosterUsernameAndCategoryTitleDescendingRow{
+		{
+			Idlinker:       1,
+			Title:          sql.NullString{String: "Example", Valid: true},
+			Url:            sql.NullString{String: "http://example.com", Valid: true},
+			Description:    sql.NullString{String: "desc", Valid: true},
+			Listed:         sql.NullTime{Time: time.Unix(0, 0), Valid: true},
+			Posterusername: sql.NullString{String: "bob", Valid: true},
+		},
+	}
+	r := httptest.NewRequest("GET", "http://example.com/linker/rss", nil)
+	feed := linkerFeed(r, rows)
+	if len(feed.Items) != 1 {
+		t.Fatalf("expected 1 item got %d", len(feed.Items))
+	}
+	if feed.Items[0].Link.Href != "http://example.com" {
+		t.Errorf("unexpected link %s", feed.Items[0].Link.Href)
+	}
+	if feed.Items[0].Title != "Example" {
+		t.Errorf("unexpected title %s", feed.Items[0].Title)
+	}
+}
 func TestLinkerApproveAddsToSearch(t *testing.T) {
 	db, mock, err := sqlmock.New()
 	if err != nil {
